+++ NEWS.md
@@ -66,6 +66,8 @@
 
 - The default theme in `new_site()` was changed from `hugo-lithium` to [`hugo-prose`](https://github.com/yihui/hugo-prose) (this may change again in the future).
 
+- When clicking the "Build Website" button in RStudio, it will no longer open the `index.html` file in the generated site folder, but emit a message telling users that this folder needs to be served before the web pages can be correctly previewed (#522).
+
 - For Jekyll sites, the arguments `--watch`, `--incremental`, and `--livereload` are passed to `jekyll serve` by default. These arguments can be set via the global R option `blogdown.jekyll.server`.
 
 - The meta variable `link-citations` is set to `true` for the `pandoc_args` argument of `blogdown::html_page()`, so that links can be generated on citation items by default.